fix: share deterministic regular file reader
