+++ tools/repo/build-data-contract-bundle.mjs
@@ -1,7 +1,7 @@
-import { constants } from "node:fs";
-import { lstat, open, readFile, writeFile } from "node:fs/promises";
+import { writeFile } from "node:fs/promises";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
+import { isSafeRelativePath, readRegularFile } from "../lib/read-regular-file.mjs";
 
 const bundlePath = "contracts/bundles/data-contracts-v1.0.0.json";
 const resources = [
@@ -40,10 +40,6 @@ const externalSourceRegistrations = [
   },
 ];
 
-const safePath = (value) => typeof value === "string" && value.length > 0 && !value.includes("\\")
-  && !path.posix.isAbsolute(value) && path.posix.normalize(value) === value
-  && value.split("/").every((part) => part !== "" && part !== "." && part !== "..");
-
 function exactKeys(value, keys, label) {
   if (!value || typeof value !== "object" || Array.isArray(value)
       || Object.keys(value).join(",") !== keys.join(",")) {
@@ -55,36 +51,13 @@ function validateClosedResources() {
   const resourceIds = new Set();
   const sourcePaths = new Set();
   for (const [resourceId, sourcePath] of resources) {
-    if (!safePath(resourceId) || !safePath(sourcePath) || resourceIds.has(resourceId) || sourcePaths.has(sourcePath)) {
+    if (!isSafeRelativePath(resourceId) || !isSafeRelativePath(sourcePath) || resourceIds.has(resourceId) || sourcePaths.has(sourcePath)) {
       throw new Error("data contract bundle resource identity is unsafe, duplicate, or unknown");
     }
     resourceIds.add(resourceId);
     sourcePaths.add(sourcePath);
   }
   if (resources.length !== 6) throw new Error("data contract bundle resource set is incomplete or unexpected");
-}
-
-async function readRegularFile(repositoryRoot, relativePath) {
-  if (!safePath(relativePath)) throw new Error("source path is unsafe");
-  const root = path.resolve(repositoryRoot);
-  const rootStat = await lstat(root);
-  if (!rootStat.isDirectory() || rootStat.isSymbolicLink()) throw new Error("repositoryRoot must be a regular directory");
-  let source = root;
-  const parts = relativePath.split("/");
-  for (const [index, part] of parts.entries()) {
-    source = path.join(source, part);
-    const stat = await lstat(source);
-    if (stat.isSymbolicLink()) throw new Error("source input must not contain a symlink");
-    if (index === parts.length - 1 && !stat.isFile()) throw new Error("source input must be a regular file");
-  }
-  const handle = await open(source, constants.O_RDONLY | constants.O_NOFOLLOW | constants.O_NONBLOCK);
-  try {
-    const stat = await handle.stat();
-    if (!stat.isFile()) throw new Error("source input must be a regular file");
-    return handle.readFile();
-  } finally {
-    await handle.close();
-  }
 }
 
 function validateProductionScope(scope) {
@@ -108,7 +81,7 @@ export async function buildDataContractBundle({ repositoryRoot }) {
   if (typeof repositoryRoot !== "string") throw new Error("repositoryRoot is required");
   validateClosedResources();
   const root = path.resolve(repositoryRoot);
-  const entries = await Promise.all(resources.map(async ([resourceId, sourcePath]) => [resourceId, await readRegularFile(root, sourcePath)]));
+  const entries = await Promise.all(resources.map(async ([resourceId, sourcePath]) => [resourceId, await readRegularFile(root, sourcePath, { label: "source input" })]));
   const scopeBytes = entries.find(([resourceId]) => resourceId === "datapack/production-datapack-scope.json")?.[1];
   if (scopeBytes == null) throw new Error("production datapack scope resource is missing");
   validateProductionScope(JSON.parse(scopeBytes));
@@ -128,7 +101,7 @@ async function main() {
   const generated = await buildDataContractBundle({ repositoryRoot: root });
   const target = path.resolve(root, bundlePath);
   if (mode === "--check") {
-    const existing = await readRegularFile(root, bundlePath);
+    const existing = await readRegularFile(root, bundlePath, { label: "data contract bundle" });
     if (!generated.equals(existing)) throw new Error("data contract bundle is not generated from current closed inputs");
     return;
   }
